Refactor module detailed page into smaller components
