+++ components/dialogs/DeleteModuleDialog.tsx
